utils/requirements.txt
pytube==15.0.0
pydub==0.25.1
DALI-dataset==1.1
tqdm==4.66.1
librosa==0.10.1
scipy==1.11.4